third_party/WebKit/LayoutTests/fast/canvas/canvas-color-serialization-expected.txt

Test that getting color properties from a CanvasRenderingContext2D returns properly formatted values.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS checkDefaultValue(ctx.strokeStyle) is '#000000'
PASS checkDefaultValue(ctx.fillStyle) is '#000000'
PASS checkDefaultValue(ctx.shadowColor) is 'rgba(0, 0, 0, 0)'
PASS trySettingStrokeStyle('transparent') is 'rgba(0, 0, 0, 0)'
PASS trySettingFillStyle('transparent') is 'rgba(0, 0, 0, 0)'
PASS trySettingShadowColor('transparent') is 'rgba(0, 0, 0, 0)'
PASS trySettingStrokeStyle('red') is '#ff0000'
PASS trySettingFillStyle('red') is '#ff0000'
PASS trySettingShadowColor('red') is '#ff0000'
PASS trySettingStrokeStyle('white') is '#ffffff'
PASS trySettingFillStyle('white') is '#ffffff'
PASS trySettingShadowColor('white') is '#ffffff'
PASS trySettingStrokeStyle('') is '#666666'
PASS trySettingFillStyle('') is '#666666'
PASS trySettingShadowColor('') is '#666666'
PASS trySettingStrokeStyle('RGBA(0, 0, 0, 0)') is 'rgba(0, 0, 0, 0)'
PASS trySettingFillStyle('RGBA(0, 0, 0, 0)') is 'rgba(0, 0, 0, 0)'
PASS trySettingShadowColor('RGBA(0, 0, 0, 0)') is 'rgba(0, 0, 0, 0)'
PASS trySettingStrokeStyle('rgba(0,255,0,1.0)') is '#00ff00'
PASS trySettingFillStyle('rgba(0,255,0,1.0)') is '#00ff00'
PASS trySettingShadowColor('rgba(0,255,0,1.0)') is '#00ff00'
PASS trySettingStrokeStyle('rgba(1,2,3,0.4)') is 'rgba(1, 2, 3, 0.4)'
PASS trySettingFillStyle('rgba(1,2,3,0.4)') is 'rgba(1, 2, 3, 0.4)'
PASS trySettingShadowColor('rgba(1,2,3,0.4)') is 'rgba(1, 2, 3, 0.4)'
PASS trySettingStrokeStyle('RgB(1,2,3)') is '#010203'
PASS trySettingFillStyle('RgB(1,2,3)') is '#010203'
PASS trySettingShadowColor('RgB(1,2,3)') is '#010203'
PASS trySettingStrokeStyle('rGbA(1,2,3,0)') is 'rgba(1, 2, 3, 0)'
PASS trySettingFillStyle('rGbA(1,2,3,0)') is 'rgba(1, 2, 3, 0)'
PASS trySettingShadowColor('rGbA(1,2,3,0)') is 'rgba(1, 2, 3, 0)'
PASS trySettingStrokeStyle(true) is '#666666'
PASS trySettingFillStyle(true) is '#666666'
PASS trySettingShadowColor(true) is '#666666'
PASS trySettingStrokeStyle(false) is '#666666'
PASS trySettingFillStyle(false) is '#666666'
PASS trySettingShadowColor(false) is '#666666'
PASS trySettingStrokeStyle(0) is '#666666'
PASS trySettingFillStyle(0) is '#666666'
PASS trySettingShadowColor(0) is '#666666'
PASS trySettingStrokeStyle(1) is '#666666'
PASS trySettingFillStyle(1) is '#666666'
PASS trySettingShadowColor(1) is '#666666'
PASS trySettingStrokeStyle(-1) is '#666666'
PASS trySettingFillStyle(-1) is '#666666'
PASS trySettingShadowColor(-1) is '#666666'
PASS trySettingStrokeStyle(NaN) is '#666666'
PASS trySettingFillStyle(NaN) is '#666666'
PASS trySettingShadowColor(NaN) is '#666666'
PASS trySettingStrokeStyle(Infinity) is '#666666'
PASS trySettingFillStyle(Infinity) is '#666666'
PASS trySettingShadowColor(Infinity) is '#666666'
PASS trySettingStrokeStyle(null) is '#666666'
PASS trySettingFillStyle(null) is '#666666'
PASS trySettingShadowColor(null) is '#666666'
PASS successfullyParsed is true

TEST COMPLETE
